BackEnd Merge

See merge request s11-webmobile1-sub2/S11P12B302!105

+++ back-end/koplaydev/src/main/java/com/edu/koplay/model/Student.java
@@ -14,13 +14,13 @@ import java.util.Date;
 @Data
 @Entity
 public class Student {
-    public Student(int studentIdx) {
+    public Student(Long studentIdx) {
         this.studentIdx = studentIdx;
     }
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int studentIdx;
+    private Long studentIdx;
 
     @ManyToOne
     @JoinColumn(name = "parent_idx", nullable = false)
